Create File for changelog type and new component
