+++ src/types/LogTasks.ts
@@ -1,0 +1,9 @@
+type Task = {
+  kind: "in-progress" | "done" | "fix";
+  content: string;
+};
+
+export type LogTasks = {
+  month: string;
+  tasks: Task[];
+};
